get rid of calimero-identity
