--- crates/identity/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "calimero-identity"
-version = "0.1.0"
-authors.workspace = true
-edition.workspace = true
-repository.workspace = true
-license.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bs58.workspace = true
-eyre.workspace = true
-hex.workspace = true
-libp2p = { workspace = true }
-web3.workspace = true
-
-calimero-primitives = { path = "../primitives" }
-
-[lints]
-workspace = true
