--- crates/identity/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod auth;
