+++ crates/server/Cargo.toml
@@ -17,6 +17,7 @@ chrono.workspace = true
 ed25519-dalek.workspace = true
 eyre.workspace = true
 futures-util = { workspace = true, optional = true }
+hex.workspace = true
 jsonwebtoken.workspace = true
 libp2p.workspace = true
 local-ip-address.workspace = true
@@ -36,11 +37,11 @@ tower.workspace = true
 tower-http = { workspace = true, features = ["cors", "fs"] }
 tower-sessions = { workspace = true, optional = true }
 tracing.workspace = true
+web3.workspace = true
 x509-parser.workspace = true
 
 calimero-context = { path = "../context" }
 calimero-node-primitives = { path = "../node-primitives" }
-calimero-identity = { path = "../identity" }
 calimero-primitives = { path = "../primitives" }
 calimero-server-primitives = { path = "../server-primitives" }
 calimero-store = { path = "../store", features = ["serde"] }
