+++ crates/server/src/admin/utils.rs
@@ -1,2 +1,4 @@
 pub mod auth;
+#[cfg(test)]
+mod auth_tests;
 pub mod jwt;
