+++ crates/server/src/admin/utils/auth_tests.rs
